+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 #Changelog
 ---
+####4.6.2
+- Stacktraces now get logged
+- More logging at while starting the bot
+
 ####4.6.1
 - Error logging improvements